Bug fixes for traffic Control * Fixed type/overflow/cutoff issues. * Fixed copy/paste bug in crash detection (and tested with forcing crash temporarily). * Fixed inconsistent if/else and missing break statements in switch: both in set*TrafficLight() functions. * Changed to faster than realtime simulation.
Still room for optimization (for lights) and refactoring for cleaner code.

+++ Debugging_Challenge/TrafficControl/traffic_light_problem.h
@@ -33,13 +33,13 @@ typedef enum
 //****************************************************************************
 struct lane_of_cars_s
 {
-	int8_t carsWaitingAtIntersection;
-	int8_t carsInIntersection;
-	int8_t carsLeavingIntersection;
-	int8_t carsThatHaveLeft;
+	uint16_t carsWaitingAtIntersection;
+	uint8_t carsInIntersection;
+	uint8_t carsLeavingIntersection;
+	uint8_t carsThatHaveLeft;
 	
-	int8_t popularity;
-	int8_t timeWaiting;
+	uint8_t popularity;
+	uint16_t timeWaiting;
 };
 
 struct intersection_s
